Se modificó archivos html, se crea función de filtrado por tipo en data.js
