+++ src/index.html
@@ -53,7 +53,7 @@
       <section id="Titulo">
         <div class="titulo-principal">
          <h1>Bienvenido a Mundo</h1>
-          <img src="./img/titulo.png" alt="imagen-titulo" >
+          <img src="./img/titulo.png" alt="imagenTitulo" >
         </div>
 
       </section>
@@ -61,7 +61,7 @@
      <div id="conteItemsCarrusel">
       <div class="itemCarrusel" id="itemCarrusel-1">
         <div class="tarjetaCarrusel" id="tarjetaCarrusel-1">
-         <img src="./img/tipos-de-pokemon.png" alt="tipos-pokemon">
+         <img src="./img/tipos-de-pokemon.png" alt="tiposPokemon">
          <div class="texto-Foto">
           <h3>Conoce tu pokemon favorito</h3>
          </div>
@@ -122,7 +122,7 @@
   <div id="root">
   </div>
     <script src="main.js" type="module"></script>
-    <script src="pokemon.js"></script>
+
   
   </body>
 </html>
