+++ src/tiposPoquemon.html
@@ -60,24 +60,25 @@
   
       <section id="seccion2">
      <div class="imagenes">
-      <img src="./img/bicho.png" alt="bicho">
-      <img src="./img/dragon.png" alt="dragon">
-      <img src="./img/electricco.png" alt="electrico">
-      <img src="./img/hada.png" alt="hada">
-      <img src="./img/lucha.png" alt="lucha">
-      <img src="./img/fuego.png" alt="fuego">
-      <img src="./img/volar.png" alt="volar">
-      <img src="./img/fantasma.png" alt="fantasma">
-      <img src="./img/planta.png" alt="planta">
-      <img src="./img/tierra.png" alt="tierra">
-      <img src="./img/hielo.png" alt="hielo">
-      <img src="./img/Normal.png" alt="Normal">
-      <img src="./img/Veneno.png" alt="Veneno">
-      <img src="./img/psiquico.png" alt="psiquico">
-      <img src="./img/roca.png" alt="roca">
-      <img src="./img/acero.jpg" alt="acero">
-      <img src="./img/agua.png" alt="agua">
+      <img src="./img/bicho.png"  class="filtro-tipo"  id="bug"  alt="bug">
+      <img src="./img/dragon.png" class="filtro-tipo" id="dragon" alt="dragon">
+      <img src="./img/electricco.png" class="filtro-tipo" id="electric" alt="electric">
+      <img src="./img/hada.png"  class="filtro-tipo" id="fairy" alt="fairy">
+      <img src="./img/lucha.png"  class="filtro-tipo" id="fighting" alt="fighting">
+      <img src="./img/fuego.png"  class="filtro-tipo" id="fire" alt="fire">
+      <img src="./img/volar.png"  class="filtro-tipo" id="flying" alt="flying">
+      <img src="./img/fantasma.png"  class="filtro-tipo" id="ghost" alt="ghost">
+      <img src="./img/planta.png"  class="filtro-tipo" id="grass" alt="grass">
+      <img src="./img/tierra.png" class="filtro-tipo" id="ground"  alt="ground">
+      <img src="./img/hielo.png" class="filtro-tipo" id="ice" alt="ice">
+      <img src="./img/Normal.png"  class="filtro-tipo" id="normal" alt="normal">
+      <img src="./img/Veneno.png"  class="filtro-tipo" id="poison" alt="poison">
+      <img src="./img/psiquico.png"  class="filtro-tipo" id="psychic" alt="psychic">
+      <img src="./img/roca.png"  class="filtro-tipo" id="rock" alt="rock">
+      <img src="./img/acero.jpg"  class="filtro-tipo" id="steel" alt="steel">
+      <img src="./img/agua.png" class="filtro-tipo" id="water" alt="water">
      </div>
+     <button id="ver-todos">Mostrar Todos</button>
       </section>
 
       <section id="seccion3">
